feat: migrate entity page store to be hooks-based
