+++ apps/web/partials/entity-page/entity-page-referenced-by.tsx
@@ -5,7 +5,7 @@ import Link from 'next/link';
 
 import * as React from 'react';
 
-import { useEntityPageStore } from '~/core/hooks/use-entity-page-store';
+import { useEntityPageStore } from '~/core/state/entity-page-store/entity-store';
 import { OmitStrict } from '~/core/types';
 import { Entity } from '~/core/utils/entity';
 import { NavUtils, getImagePath } from '~/core/utils/utils';
