+++ apps/web/partials/entity-page/readable-entity-page.tsx
@@ -1,7 +1,7 @@
 import { SYSTEM_IDS } from '@geogenesis/ids';
 
 import { useActionsStore } from '~/core/hooks/use-actions-store';
-import { useEntityPageStore } from '~/core/hooks/use-entity-page-store';
+import { useEntityPageStore } from '~/core/state/entity-page-store/entity-store';
 import { Triple } from '~/core/types';
 import { NavUtils, groupBy } from '~/core/utils/utils';
